@@ -3,10 +3,10 @@ name: crashlytics
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
+  firebase_core: ^1.1.0
+  firebase_crashlytics: ^2.0.2
   flutter:
     sdk: flutter
-  firebase_core: ^0.7.0
-  firebase_crashlytics: ^0.4.0+1
