@@ -3,21 +3,22 @@ name: fast_shopping_bloc
 publish_to: 'none'
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
-  bloc: ^6.1.0
-  clock: ^1.0.1
-  freezed_annotation: ^0.12.0
-  json_annotation: ^3.1.1
-  uuid: ^2.0.4
+  bloc: ^7.0.0
+  clock: ^1.1.0
+  collection: ^1.15.0
+  freezed_annotation: ^0.14.1
+  json_annotation: ^4.0.1
+  uuid: ^3.0.4
 
 dev_dependencies:
-  bloc_test: ^7.1.0
-  build_runner: ^1.7.3
-  freezed: ^0.12.3
-  json_serializable: ^3.5.1
-  lint: ^1.2.0
-  mockito: ^4.1.3
-  test: ^1.15.7
-  test_coverage: ^0.5.0
+  bloc_test: ^8.0.0
+  build_runner: ^2.0.2
+  coverage: ^1.0.2
+  freezed: ^0.14.1+3
+  json_serializable: ^4.1.1
+  lint: ^1.5.3
+  mockito: ^5.0.7
+  test: ^1.17.3
