@@ -6,39 +6,44 @@ publish_to: 'none'
 version: 3.0.4+23
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: '>=2.2.0-0 < 3.0.0'
 
 dependencies:
-  flutter:
-    sdk: flutter
-  flutter_localizations:
-    sdk: flutter
-  animations: ^1.0.0+5
-  clock: ^1.0.1
+  animations: ^2.0.0
+  clock: ^1.1.0
   crashlytics:
     path: crashlytics/foss # CD changes me 🤖
   fast_shopping_bloc:
     path: fast_shopping_bloc
-  flutter_hooks: ^0.15.0
-  flutter_bloc: ^7.0.0-dev.2
-  flutter_svg: ^0.19.0
-  implicitly_animated_reorderable_list: ^0.4.0-dev+3
-  intl: ^0.16.1
-  md2_tab_indicator: ^1.0.2
-  shared_preferences: ^0.5.6+1
-  timeago: ^2.0.22
-  url_launcher: ^5.4.2
-  uuid: ^2.0.4
+  flutter:
+    sdk: flutter
+  flutter_bloc: ^7.0.0
+  flutter_hooks: ^0.16.0
+  flutter_localizations:
+    sdk: flutter
+  flutter_svg: ^0.22.0
+  implicitly_animated_reorderable_list: ^0.4.0
+  intl: ^0.17.0
+  md2_tab_indicator:
+    git:
+      url: https://github.com/westdabestdb/md2_tab_indicator.git
+      ref: 1097deb4f4018c408fd4745711efb8f0ea9044e7
+  shared_preferences: ^2.0.5
+  timeago: ^3.0.2
+  url_launcher: ^6.0.3
+  uuid: ^3.0.4
 
 dev_dependencies:
-  flutter_test:
-    sdk: flutter
+  build_runner: ^2.0.2
   flutter_driver:
     sdk: flutter
-  image: ^2.1.19
-  lint: ^1.2.0
-  mockito: ^4.1.3
-  test: ^1.15.5
+  flutter_test:
+    sdk: flutter
+  image: ^3.0.2
+  lint: ^1.5.3
+  mockito: ^5.0.7
+  test: ^1.16.8
 
 flutter:
   generate: true
