@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'shopping_lists_cubit.dart';
 
@@ -9,13 +9,15 @@ part of 'shopping_lists_cubit.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$ShoppingListsStateTearOff {
   const _$ShoppingListsStateTearOff();
 
-// ignore: unused_element
   _ShoppingListsState call(
-      {String selectedId, List<ShoppingList> lists = const []}) {
+      {String? selectedId, List<ShoppingList> lists = const []}) {
     return _ShoppingListsState(
       selectedId: selectedId,
       lists: lists,
@@ -24,16 +26,16 @@ class _$ShoppingListsStateTearOff {
 }
 
 /// @nodoc
-// ignore: unused_element
 const $ShoppingListsState = _$ShoppingListsStateTearOff();
 
 /// @nodoc
 mixin _$ShoppingListsState {
-  String get selectedId;
-  List<ShoppingList> get lists;
+  String? get selectedId => throw _privateConstructorUsedError;
+  List<ShoppingList> get lists => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $ShoppingListsStateCopyWith<ShoppingListsState> get copyWith;
+  $ShoppingListsStateCopyWith<ShoppingListsState> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -41,7 +43,7 @@ abstract class $ShoppingListsStateCopyWith<$Res> {
   factory $ShoppingListsStateCopyWith(
           ShoppingListsState value, $Res Function(ShoppingListsState) then) =
       _$ShoppingListsStateCopyWithImpl<$Res>;
-  $Res call({String selectedId, List<ShoppingList> lists});
+  $Res call({String? selectedId, List<ShoppingList> lists});
 }
 
 /// @nodoc
@@ -55,13 +57,18 @@ class _$ShoppingListsStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object selectedId = freezed,
-    Object lists = freezed,
+    Object? selectedId = freezed,
+    Object? lists = freezed,
   }) {
     return _then(_value.copyWith(
-      selectedId:
-          selectedId == freezed ? _value.selectedId : selectedId as String,
-      lists: lists == freezed ? _value.lists : lists as List<ShoppingList>,
+      selectedId: selectedId == freezed
+          ? _value.selectedId
+          : selectedId // ignore: cast_nullable_to_non_nullable
+              as String?,
+      lists: lists == freezed
+          ? _value.lists
+          : lists // ignore: cast_nullable_to_non_nullable
+              as List<ShoppingList>,
     ));
   }
 }
@@ -73,7 +80,7 @@ abstract class _$ShoppingListsStateCopyWith<$Res>
           _ShoppingListsState value, $Res Function(_ShoppingListsState) then) =
       __$ShoppingListsStateCopyWithImpl<$Res>;
   @override
-  $Res call({String selectedId, List<ShoppingList> lists});
+  $Res call({String? selectedId, List<ShoppingList> lists});
 }
 
 /// @nodoc
@@ -89,71 +96,36 @@ class __$ShoppingListsStateCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object selectedId = freezed,
-    Object lists = freezed,
+    Object? selectedId = freezed,
+    Object? lists = freezed,
   }) {
     return _then(_ShoppingListsState(
-      selectedId:
-          selectedId == freezed ? _value.selectedId : selectedId as String,
-      lists: lists == freezed ? _value.lists : lists as List<ShoppingList>,
+      selectedId: selectedId == freezed
+          ? _value.selectedId
+          : selectedId // ignore: cast_nullable_to_non_nullable
+              as String?,
+      lists: lists == freezed
+          ? _value.lists
+          : lists // ignore: cast_nullable_to_non_nullable
+              as List<ShoppingList>,
     ));
   }
 }
 
 /// @nodoc
+
 class _$_ShoppingListsState extends _ShoppingListsState {
-  _$_ShoppingListsState({this.selectedId, this.lists = const []})
-      : assert(lists != null),
-        super._();
+  _$_ShoppingListsState({this.selectedId, this.lists = const []}) : super._();
 
   @override
-  final String selectedId;
+  final String? selectedId;
   @JsonKey(defaultValue: const [])
   @override
   final List<ShoppingList> lists;
 
-  bool _didselected = false;
-  ShoppingList _selected;
-
-  @override
-  ShoppingList get selected {
-    if (_didselected == false) {
-      _didselected = true;
-      _selected = lists.firstWhere(
-        (list) => list.id == selectedId,
-        orElse: () => null,
-      );
-    }
-    return _selected;
-  }
-
-  bool _didcurrent = false;
-  List<ShoppingList> _current;
-
-  @override
-  List<ShoppingList> get current {
-    if (_didcurrent == false) {
-      _didcurrent = true;
-      _current = lists.where((list) => list.archivedAt == null).toList();
-    }
-    return _current;
-  }
-
-  bool _didarchived = false;
-  List<ShoppingList> _archived;
-
-  @override
-  List<ShoppingList> get archived {
-    if (_didarchived == false) {
-      _didarchived = true;
-      _archived = lists.where((list) => list.archivedAt != null).toList();
-    }
-    return _archived;
-  }
-
   @override
   String toString() {
-    return 'ShoppingListsState(selectedId: $selectedId, lists: $lists, selected: $selected, current: $current, archived: $archived)';
+    return 'ShoppingListsState(selectedId: $selectedId, lists: $lists)';
   }
 
   @override
@@ -180,15 +152,16 @@ class _$_ShoppingListsState extends _ShoppingListsState {
 }
 
 abstract class _ShoppingListsState extends ShoppingListsState {
-  _ShoppingListsState._() : super._();
-  factory _ShoppingListsState({String selectedId, List<ShoppingList> lists}) =
+  factory _ShoppingListsState({String? selectedId, List<ShoppingList> lists}) =
       _$_ShoppingListsState;
+  _ShoppingListsState._() : super._();
 
   @override
-  String get selectedId;
+  String? get selectedId => throw _privateConstructorUsedError;
   @override
-  List<ShoppingList> get lists;
+  List<ShoppingList> get lists => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$ShoppingListsStateCopyWith<_ShoppingListsState> get copyWith;
+  _$ShoppingListsStateCopyWith<_ShoppingListsState> get copyWith =>
+      throw _privateConstructorUsedError;
 }
