@@ -3,7 +3,7 @@ import 'package:fast_shopping/theme.dart';
 import 'package:flutter/material.dart';
 
 class RemoveAllDoneDialog extends StatelessWidget {
-  const RemoveAllDoneDialog({Key key, @required this.itemsCount})
+  const RemoveAllDoneDialog({Key? key, required this.itemsCount})
       : super(key: key);
 
   final int itemsCount;
@@ -12,16 +12,16 @@ class RemoveAllDoneDialog extends StatelessWidget {
   Widget build(BuildContext context) {
     return AlertDialog(
       backgroundColor: Theme.of(context).colorScheme.surface,
-      title: Text(S.of(context).remove_all_done_dialog_title),
-      content: Text(S.of(context).remove_all_done_dialog_body(itemsCount)),
+      title: Text(S.of(context)!.remove_all_done_dialog_title),
+      content: Text(S.of(context)!.remove_all_done_dialog_body(itemsCount)),
       actions: [
         TextButton(
           onPressed: () => Navigator.of(context).pop(false),
-          child: Text(S.of(context).remove_all_done_dialog_cancel),
+          child: Text(S.of(context)!.remove_all_done_dialog_cancel),
         ),
         TextButton(
           onPressed: () => Navigator.of(context).pop(true),
-          child: Text(S.of(context).remove_all_done_dialog_remove),
+          child: Text(S.of(context)!.remove_all_done_dialog_remove),
         ).danger,
       ],
     );
