@@ -3,7 +3,7 @@ import 'package:fast_shopping/l10n/l10n.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class AppLogo extends StatelessWidget {
-  const AppLogo({Key key}) : super(key: key);
+  const AppLogo({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -18,7 +18,7 @@ class AppLogo extends StatelessWidget {
         ),
         const SizedBox(width: 16),
         Text(
-          S.of(context).app_title,
+          S.of(context)!.app_title,
           style: Theme.of(context).textTheme.headline6,
         ),
       ],
