@@ -7,11 +7,11 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 
 class ContentSpace extends HookWidget {
   const ContentSpace({
-    Key key,
-    @required this.titleController,
-    @required this.item,
-    @required this.editing,
-    @required this.onEditingChanged,
+    Key? key,
+    required this.titleController,
+    required this.item,
+    required this.editing,
+    required this.onEditingChanged,
     this.onTitleChanged,
   }) : super(key: key);
 
@@ -20,14 +20,14 @@ class ContentSpace extends HookWidget {
   final Item item;
   final bool editing;
   final ValueChanged<bool> onEditingChanged;
-  final ValueChanged<String> onTitleChanged;
+  final ValueChanged<String>? onTitleChanged;
 
   VoidCallback Function() _onItemOrEditingUpdate(
     BuildContext context,
     FocusScopeNode focusScopeNode,
     FocusNode focusNode,
   ) {
-    final debounce = useState<Timer>(null);
+    final debounce = useState<Timer?>(null);
 
     return () {
       // Focus or unfocus title field based on the item tile state
@@ -44,7 +44,8 @@ class ContentSpace extends HookWidget {
       //     editing item -> not editing item with new title
       // TODO: Add debouncing (ideally conditional) to the SelectedShoppingListCubit
       //       after migrating to bloc 7.0.0 (transformTransitions in cubits).
-      if (debounce.value?.isActive ?? false) debounce.value.cancel();
+      // Reply: If I only could recall what exactly I wanted to debounce then...
+      if (debounce.value?.isActive ?? false) debounce.value!.cancel();
       debounce.value = Timer(const Duration(milliseconds: 50), () {
         titleController.value = TextEditingValue(
           text: item.title,
@@ -88,8 +89,8 @@ class ContentSpace extends HookWidget {
               const EdgeInsets.symmetric(horizontal: 32, vertical: 14),
           isDense: true,
           hintText: editing
-              ? S.of(context).list_item_title_hint
-              : S.of(context).list_item_no_name,
+              ? S.of(context)!.list_item_title_hint
+              : S.of(context)!.list_item_no_name,
           hintStyle: TextStyle(
             fontStyle: editing ? FontStyle.normal : FontStyle.italic,
           ),
