@@ -5,8 +5,8 @@ import 'package:fast_shopping/l10n/l10n.dart';
 
 class DeleteListDialog extends StatelessWidget {
   const DeleteListDialog({
-    Key key,
-    @required this.shoppingList,
+    Key? key,
+    required this.shoppingList,
   }) : super(key: key);
 
   final ShoppingList shoppingList;
@@ -15,13 +15,13 @@ class DeleteListDialog extends StatelessWidget {
   Widget build(BuildContext context) {
     const divider = r'$$DIVIDER$$';
     final bodyParts =
-        S.of(context).delete_shopping_list_dialog_body(divider).split(divider);
+        S.of(context)!.delete_shopping_list_dialog_body(divider).split(divider);
     final bodyBefore = bodyParts.first;
     final bodyAfter = bodyParts.last;
 
     return AlertDialog(
       backgroundColor: Theme.of(context).colorScheme.surface,
-      title: Text(S.of(context).delete_shopping_list_dialog_title),
+      title: Text(S.of(context)!.delete_shopping_list_dialog_title),
       content: Text.rich(TextSpan(children: [
         TextSpan(text: bodyBefore),
         TextSpan(
@@ -33,11 +33,11 @@ class DeleteListDialog extends StatelessWidget {
       actions: [
         TextButton(
           onPressed: () => Navigator.of(context).pop(false),
-          child: Text(S.of(context).delete_shopping_list_dialog_cancel),
+          child: Text(S.of(context)!.delete_shopping_list_dialog_cancel),
         ),
         TextButton(
           onPressed: () => Navigator.of(context).pop(true),
-          child: Text(S.of(context).delete_shopping_list_dialog_delete),
+          child: Text(S.of(context)!.delete_shopping_list_dialog_delete),
         ).danger,
       ],
     );
