@@ -3,15 +3,15 @@ import 'package:flutter/material.dart';
 import 'package:fast_shopping/l10n/l10n.dart';
 
 class NewListDialog extends StatelessWidget {
-  const NewListDialog({Key key}) : super(key: key);
+  const NewListDialog({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return PromptDialog(
-      title: S.of(context).add_list_dialog_title,
-      inputHint: S.of(context).add_list_dialog_name_hint,
-      submitText: S.of(context).add_list_dialog_add,
-      cancelText: S.of(context).add_list_dialog_cancel,
+      title: S.of(context)!.add_list_dialog_title,
+      inputHint: S.of(context)!.add_list_dialog_name_hint,
+      submitText: S.of(context)!.add_list_dialog_add,
+      cancelText: S.of(context)!.add_list_dialog_cancel,
     );
   }
 }
