@@ -2,12 +2,12 @@ import 'package:flutter/material.dart';
 import 'package:fast_shopping/l10n/l10n.dart';
 
 class NoArchivedListsPlaceholder extends StatelessWidget {
-  const NoArchivedListsPlaceholder({Key key}) : super(key: key);
+  const NoArchivedListsPlaceholder({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     return Center(
-      child: Text(S.of(context).no_archived_lists_message),
+      child: Text(S.of(context)!.no_archived_lists_message),
     );
   }
 }
