@@ -3,7 +3,7 @@ import 'package:fast_shopping/l10n/l10n.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 
 class CurrentTabPlaceholder extends StatelessWidget {
-  const CurrentTabPlaceholder({Key key}) : super(key: key);
+  const CurrentTabPlaceholder({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
@@ -12,7 +12,7 @@ class CurrentTabPlaceholder extends StatelessWidget {
       children: [
         SvgPicture.asset('assets/shopping_bags_woman.svg', width: 250),
         const SizedBox(height: 32),
-        Text(S.of(context).no_current_lists_message),
+        Text(S.of(context)!.no_current_lists_message),
       ],
     );
   }
