@@ -5,7 +5,7 @@ import 'shopping_lists_tab_bar.dart';
 
 class ShoppingListsAppBar extends StatelessWidget
     implements PreferredSizeWidget {
-  const ShoppingListsAppBar({Key key, @required this.tabController})
+  const ShoppingListsAppBar({Key? key, required this.tabController})
       : super(key: key);
 
   final TabController tabController;
@@ -14,7 +14,7 @@ class ShoppingListsAppBar extends StatelessWidget
   Widget build(BuildContext context) {
     return AppBar(
       leading: const BackButton(key: ValueKey('lists-back')),
-      title: Text(S.of(context).shopping_lists_title),
+      title: Text(S.of(context)!.shopping_lists_title),
       bottom: ShoppingListsTabBar(controller: tabController),
     );
   }
