@@ -1,5 +1,5 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
-// ignore_for_file: deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides
 
 part of 'app_settings_cubit.dart';
 
@@ -9,38 +9,37 @@ part of 'app_settings_cubit.dart';
 
 T _$identity<T>(T value) => value;
 
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more informations: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
 /// @nodoc
 class _$AppSettingsTearOff {
   const _$AppSettingsTearOff();
 
-// ignore: unused_element
   _AppSettings call(
       {ShoppingListsMode shoppingListsMode = ShoppingListsMode.multiple,
       DarkTheme darkTheme = DarkTheme.system,
-      ItemsLayout itemsLayout = ItemsLayout.comfortable,
       bool moveDoneToEnd = false}) {
     return _AppSettings(
       shoppingListsMode: shoppingListsMode,
       darkTheme: darkTheme,
-      itemsLayout: itemsLayout,
       moveDoneToEnd: moveDoneToEnd,
     );
   }
 }
 
 /// @nodoc
-// ignore: unused_element
 const $AppSettings = _$AppSettingsTearOff();
 
 /// @nodoc
 mixin _$AppSettings {
-  ShoppingListsMode get shoppingListsMode;
-  DarkTheme get darkTheme;
-  ItemsLayout get itemsLayout;
-  bool get moveDoneToEnd;
+  ShoppingListsMode get shoppingListsMode => throw _privateConstructorUsedError;
+  DarkTheme get darkTheme => throw _privateConstructorUsedError;
+  bool get moveDoneToEnd => throw _privateConstructorUsedError;
 
   @JsonKey(ignore: true)
-  $AppSettingsCopyWith<AppSettings> get copyWith;
+  $AppSettingsCopyWith<AppSettings> get copyWith =>
+      throw _privateConstructorUsedError;
 }
 
 /// @nodoc
@@ -51,7 +50,6 @@ abstract class $AppSettingsCopyWith<$Res> {
   $Res call(
       {ShoppingListsMode shoppingListsMode,
       DarkTheme darkTheme,
-      ItemsLayout itemsLayout,
       bool moveDoneToEnd});
 }
 
@@ -65,23 +63,23 @@ class _$AppSettingsCopyWithImpl<$Res> implements $AppSettingsCopyWith<$Res> {
 
   @override
   $Res call({
-    Object shoppingListsMode = freezed,
-    Object darkTheme = freezed,
-    Object itemsLayout = freezed,
-    Object moveDoneToEnd = freezed,
+    Object? shoppingListsMode = freezed,
+    Object? darkTheme = freezed,
+    Object? moveDoneToEnd = freezed,
   }) {
     return _then(_value.copyWith(
       shoppingListsMode: shoppingListsMode == freezed
           ? _value.shoppingListsMode
-          : shoppingListsMode as ShoppingListsMode,
-      darkTheme:
-          darkTheme == freezed ? _value.darkTheme : darkTheme as DarkTheme,
-      itemsLayout: itemsLayout == freezed
-          ? _value.itemsLayout
-          : itemsLayout as ItemsLayout,
+          : shoppingListsMode // ignore: cast_nullable_to_non_nullable
+              as ShoppingListsMode,
+      darkTheme: darkTheme == freezed
+          ? _value.darkTheme
+          : darkTheme // ignore: cast_nullable_to_non_nullable
+              as DarkTheme,
       moveDoneToEnd: moveDoneToEnd == freezed
           ? _value.moveDoneToEnd
-          : moveDoneToEnd as bool,
+          : moveDoneToEnd // ignore: cast_nullable_to_non_nullable
+              as bool,
     ));
   }
 }
@@ -96,7 +94,6 @@ abstract class _$AppSettingsCopyWith<$Res>
   $Res call(
       {ShoppingListsMode shoppingListsMode,
       DarkTheme darkTheme,
-      ItemsLayout itemsLayout,
       bool moveDoneToEnd});
 }
 
@@ -112,38 +109,34 @@ class __$AppSettingsCopyWithImpl<$Res> extends _$AppSettingsCopyWithImpl<$Res>
 
   @override
   $Res call({
-    Object shoppingListsMode = freezed,
-    Object darkTheme = freezed,
-    Object itemsLayout = freezed,
-    Object moveDoneToEnd = freezed,
+    Object? shoppingListsMode = freezed,
+    Object? darkTheme = freezed,
+    Object? moveDoneToEnd = freezed,
   }) {
     return _then(_AppSettings(
       shoppingListsMode: shoppingListsMode == freezed
           ? _value.shoppingListsMode
-          : shoppingListsMode as ShoppingListsMode,
-      darkTheme:
-          darkTheme == freezed ? _value.darkTheme : darkTheme as DarkTheme,
-      itemsLayout: itemsLayout == freezed
-          ? _value.itemsLayout
-          : itemsLayout as ItemsLayout,
+          : shoppingListsMode // ignore: cast_nullable_to_non_nullable
+              as ShoppingListsMode,
+      darkTheme: darkTheme == freezed
+          ? _value.darkTheme
+          : darkTheme // ignore: cast_nullable_to_non_nullable
+              as DarkTheme,
       moveDoneToEnd: moveDoneToEnd == freezed
           ? _value.moveDoneToEnd
-          : moveDoneToEnd as bool,
+          : moveDoneToEnd // ignore: cast_nullable_to_non_nullable
+              as bool,
     ));
   }
 }
 
 /// @nodoc
+
 class _$_AppSettings implements _AppSettings {
   const _$_AppSettings(
       {this.shoppingListsMode = ShoppingListsMode.multiple,
       this.darkTheme = DarkTheme.system,
-      this.itemsLayout = ItemsLayout.comfortable,
-      this.moveDoneToEnd = false})
-      : assert(shoppingListsMode != null),
-        assert(darkTheme != null),
-        assert(itemsLayout != null),
-        assert(moveDoneToEnd != null);
+      this.moveDoneToEnd = false});
 
   @JsonKey(defaultValue: ShoppingListsMode.multiple)
   @override
@@ -151,16 +144,13 @@ class _$_AppSettings implements _AppSettings {
   @JsonKey(defaultValue: DarkTheme.system)
   @override
   final DarkTheme darkTheme;
-  @JsonKey(defaultValue: ItemsLayout.comfortable)
-  @override
-  final ItemsLayout itemsLayout;
   @JsonKey(defaultValue: false)
   @override
   final bool moveDoneToEnd;
 
   @override
   String toString() {
-    return 'AppSettings(shoppingListsMode: $shoppingListsMode, darkTheme: $darkTheme, itemsLayout: $itemsLayout, moveDoneToEnd: $moveDoneToEnd)';
+    return 'AppSettings(shoppingListsMode: $shoppingListsMode, darkTheme: $darkTheme, moveDoneToEnd: $moveDoneToEnd)';
   }
 
   @override
@@ -173,9 +163,6 @@ class _$_AppSettings implements _AppSettings {
             (identical(other.darkTheme, darkTheme) ||
                 const DeepCollectionEquality()
                     .equals(other.darkTheme, darkTheme)) &&
-            (identical(other.itemsLayout, itemsLayout) ||
-                const DeepCollectionEquality()
-                    .equals(other.itemsLayout, itemsLayout)) &&
             (identical(other.moveDoneToEnd, moveDoneToEnd) ||
                 const DeepCollectionEquality()
                     .equals(other.moveDoneToEnd, moveDoneToEnd)));
@@ -186,7 +173,6 @@ class _$_AppSettings implements _AppSettings {
       runtimeType.hashCode ^
       const DeepCollectionEquality().hash(shoppingListsMode) ^
       const DeepCollectionEquality().hash(darkTheme) ^
-      const DeepCollectionEquality().hash(itemsLayout) ^
       const DeepCollectionEquality().hash(moveDoneToEnd);
 
   @JsonKey(ignore: true)
@@ -199,18 +185,16 @@ abstract class _AppSettings implements AppSettings {
   const factory _AppSettings(
       {ShoppingListsMode shoppingListsMode,
       DarkTheme darkTheme,
-      ItemsLayout itemsLayout,
       bool moveDoneToEnd}) = _$_AppSettings;
 
   @override
-  ShoppingListsMode get shoppingListsMode;
+  ShoppingListsMode get shoppingListsMode => throw _privateConstructorUsedError;
   @override
-  DarkTheme get darkTheme;
+  DarkTheme get darkTheme => throw _privateConstructorUsedError;
   @override
-  ItemsLayout get itemsLayout;
-  @override
-  bool get moveDoneToEnd;
+  bool get moveDoneToEnd => throw _privateConstructorUsedError;
   @override
   @JsonKey(ignore: true)
-  _$AppSettingsCopyWith<_AppSettings> get copyWith;
+  _$AppSettingsCopyWith<_AppSettings> get copyWith =>
+      throw _privateConstructorUsedError;
 }
