@@ -10,16 +10,13 @@ _$_ShoppingList _$_$_ShoppingListFromJson(Map<String, dynamic> json) {
   return _$_ShoppingList(
     id: json['id'] as String,
     name: json['name'] as String,
-    createdAt: json['createdAt'] == null
-        ? null
-        : DateTime.parse(json['createdAt'] as String),
+    createdAt: DateTime.parse(json['createdAt'] as String),
     archivedAt: json['archivedAt'] == null
         ? null
         : DateTime.parse(json['archivedAt'] as String),
-    items: (json['items'] as List)
-            ?.map((e) =>
-                e == null ? null : Item.fromJson(e as Map<String, dynamic>))
-            ?.toList() ??
+    items: (json['items'] as List<dynamic>?)
+            ?.map((e) => Item.fromJson(e as Map<String, dynamic>))
+            .toList() ??
         [],
   );
 }
@@ -28,7 +25,7 @@ Map<String, dynamic> _$_$_ShoppingListToJson(_$_ShoppingList instance) =>
     <String, dynamic>{
       'id': instance.id,
       'name': instance.name,
-      'createdAt': instance.createdAt?.toIso8601String(),
+      'createdAt': instance.createdAt.toIso8601String(),
       'archivedAt': instance.archivedAt?.toIso8601String(),
       'items': instance.items,
     };
