@@ -1,5 +1,8 @@
 import 'package:fast_shopping_bloc/fast_shopping_bloc.dart';
-import 'package:mockito/mockito.dart';
+import 'package:mockito/annotations.dart';
 
-class MockShoppingListRepository extends Mock
-    implements ShoppingListRepository {}
+export 'mocks.mocks.dart';
+
+@GenerateMocks([ShoppingListRepository])
+// ignore: unused_element
+class _Mocks {}
