@@ -3,6 +3,7 @@ import 'dart:ui';
 import 'package:fast_shopping/app.dart';
 import 'package:fast_shopping/l10n/override_locale.dart';
 import 'package:fast_shopping_bloc/fast_shopping_bloc.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_driver/driver_extension.dart';
 import 'package:fast_shopping/l10n/l10n.dart';
@@ -11,8 +12,8 @@ import 'package:intl/locale.dart' as intl;
 void main() {
   enableFlutterDriverExtension(
     handler: (payload) {
-      if (payload.startsWith('locale-')) {
-        final localeString = payload.split('locale-').last;
+      if (payload?.startsWith('locale-') ?? false) {
+        final localeString = payload!.split('locale-').last;
         final locale = intl.Locale.parse(localeString);
         final dartLocale = Locale.fromSubtags(
           languageCode: locale.languageCode,
@@ -23,7 +24,7 @@ void main() {
         _main(dartLocale);
       }
 
-      return null;
+      return SynchronousFuture('');
     },
   );
 }
@@ -43,24 +44,19 @@ void _main(Locale locale) {
 
 class NullAppSettingsRepository extends AppSettingsRepository {
   @override
-  Future<ShoppingListsMode> getShoppingListsMode() async => null;
+  Future<ShoppingListsMode?> getShoppingListsMode() async => null;
   @override
   Future<void> saveShoppingListsMode(
     ShoppingListsMode shoppingListsMode,
   ) async {}
 
   @override
-  Future<DarkTheme> getDarkTheme() async => null;
+  Future<DarkTheme?> getDarkTheme() async => null;
   @override
   Future<void> saveDarkTheme(DarkTheme darkTheme) async {}
 
   @override
-  Future<ItemsLayout> getItemsLayout() async => null;
-  @override
-  Future<void> saveItemsLayout(ItemsLayout itemsLayout) async {}
-
-  @override
-  Future<bool> getMoveDoneToEnd() async => null;
+  Future<bool?> getMoveDoneToEnd() async => null;
   @override
   Future<void> saveMoveDoneToEnd(bool moveDoneToEnd) async {}
 }
@@ -145,5 +141,5 @@ class ScreenshotsShoppingListRepository extends ShoppingListRepository {
   Future<void> saveLists(List<ShoppingList> lists) async {}
 
   @override
-  Future<void> saveSelectedListId(String id) async {}
+  Future<void> saveSelectedListId(String? id) async {}
 }
